Add initial take on PowerShell implementation
This script is meant to scan a Seafile library for symlinks, create
`*.seaflnk` placeholder files for any symlinks, create symlinks from
existing placeholders, and add all symlinks to `seafile-ignore.txt`,
not necessarily in that order.

This is currently untested, by which I mean that the internal script
logic works great, but I'm not sure if Seafile will actually ignore
symlinks pointing to folders.

The intended workflow is to disable auto-sync in Seafile, create the
symlinks, run `seafile-symlink.ps1`, and re-enable autosync.

In order to permanently delete a symlinked item, you'll need to remove
both it and its `seaflnk` file. Otherwise, one or the other will be
recreated the next time this script is run.

There are some TODO's here, both in comments and holistically. I'd
like to make this script runnable from outside the library's root
directory. Later, I'd like to implement this in bash as well.